atlas: add a *VERY* basic atlas generator. Credit gimli.
For now it takes a single argument, the media/ dir containing images. It will
write media-2048-*.png and Textures.xml (both hardcoded) to that same path.

This will need a complete rework as it is not very efficient and relies on
DevIL. It will need to use sdl_image instead to avoid adding new build-deps.