atlas: finally settled on an xml schema
This is the format that TexturePacker (texturepacker.com) uses, which includes
many options like rotation, trimming, resizing, heuristic masks, etc. Sadly it
is non-free so we will have to rewrite the features.

This allows us to solve the chicken/egg problem by putting off writing our own
packer until we've settled on an unpacking feature-set.

In its most basic form, the format looks like this:

<TextureAtlas imagePath="TextureAtlas-2048-0.png" width="2048" height="2048">
    <sprite n="ArrowDown.png" x="484" y="2032" w="16" h="16" />
    <sprite n="ArrowUp.png" x="2031" y="496" w="16" h="16" />
    ...
</TextureAtlas>